pyproject.toml
[tool.poetry]
name = "drivezone"
version = "0.1.0"
description = ""
authors = ["Yurkes2020 <yurkes2020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.1"
requests = "^2.32.5"
celery = "^5.5.3"
redis = "^6.4.0"
django-celery-beat = "^2.8.1"
dj-database-url = "^3.0.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
